update_2.0
Added PostgresHook for database connection
Added comments
Task Instances renamed
